% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{bar}
\alias{bar}
\title{Interval with different densities close to endpoints}
\format{
A matrix.
}
\description{
A small, 1-dimensional data set that is symmetric about its
endpoints in terms of neighborhoods but not in terms of balls.
}
\keyword{datasets}
