% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_aux.R
\name{.BinaryBrierScore}
\alias{.BinaryBrierScore}
\title{Binary Brier Score}
\usage{
.BinaryBrierScore(predictions, time, status, tau, cause)
}
\arguments{
\item{predictions}{Numeric vector of model predictions.}

\item{time}{Numeric vector describing the time to the event of interest or
censoring.}

\item{status}{Vector of censoring status.}

\item{tau}{Time Brier score is evaluated.}

\item{cause}{Event of interest.}
}
\description{
Computes the Binary Brier Score (BBS) for binary outcomes at a specified
time point \eqn{\tau}.
}
\details{
The BBS is a measure of the accuracy of probabilistic predictions for binary
outcomes, where the predictions are the predicted probabilities of the event
of interest occurring by time \eqn{\tau}. The BBS is defined as the mean
squared difference between the predicted probabilities and the true outcome.
}
