% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_lcmm.R
\name{.fit_lcmm}
\alias{.fit_lcmm}
\title{Fits an LCMM model}
\usage{
.fit_lcmm(formula, data, mixture, subject, ng, ...)
}
\arguments{
\item{formula}{Two-sided linear formula for the fixed effects in the LCMM.}

\item{data}{Data frame with data}

\item{mixture}{One-sided formula specifying the class-specific fixed effects.}

\item{subject}{Name of the column indicating individual ids in data}

\item{ng}{Number of clusters in the LCMM model}

\item{...}{Additional arguments passed to the \code{\link[lcmm]{lcmm}}
function.}
}
\value{
An object of class lcmm
}
\description{
Fits an LCMM model
}
\seealso{
\code{\link[lcmm:lcmm]{lcmm::lcmm()}}
}
