% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.r
\name{extension}
\alias{extension}
\title{Generate and validate an extension vector for landmark parameters}
\usage{
extension(mult = 1, add = 0)
}
\arguments{
\item{mult}{multiplicative extension factor; added to \code{1} and multiplied by
the parameter.}

\item{add}{additive extension factor; added to parameter after
multiplication.}
}
\description{
This is a convenience function, modeled after
\code{\link[ggplot2:expansion]{ggplot2::expansion()}}, for generating extension vectors for the \verb{extend_*}
parameters of \code{\link[=landmarks_maxmin]{landmarks_maxmin()}} and \code{\link[=landmarks_lastfirst]{landmarks_lastfirst()}}.
}
\details{
The parameter \code{num} is extended to \code{num * (1 + extend_num[[1L]]) + extend_num[[2L]]}, and other parameters are extended analogously.

These extensions are used to generate covers whose sets are more numerous
or larger than strictly necessary to cover the finite metric space. When
performing topological analysis, e.g. constructing a nerve complex,
redundant overlap may be necessary to capture essential features of the
data.
}
