% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.R
\name{fit_longitudinal}
\alias{fit_longitudinal}
\title{Fits the specified longitudinal model for time-varying covariates up to
the landmark times}
\usage{
fit_longitudinal(
  x,
  landmarks,
  method,
  formula,
  dynamic_covariates,
  validation_fold = 0,
  cores = getOption("Ncpus", 1L),
  .warn_when_prop_few_obs = 0.25,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{landmarks}{A vector of Landmark times.}

\item{method}{Either \code{"lcmm"} or \code{"lme4"} or a function for fitting
a longitudinal data model, where the first argument is a formula, and also
has a \code{data} argument. Only needed for fit-based prediction methods
used later in \code{\link{predict_longitudinal}} (e.g. \code{"lcmm"},
\code{"lme4"}); summary measures such as \code{"locf"} are computed
directly from the data and do not require a call to
\code{fit_longitudinal}.}

\item{formula}{A formula to be used in longitudinal sub-model fitting.}

\item{dynamic_covariates}{Vector of time-varying covariates to be modelled
as the outcome of a longitudinal model.}

\item{validation_fold}{If positive, cross-validation fold where model is
fitted. If 0 (default), model fitting is performed using the complete
dataset.}

\item{cores}{Number of cores/threads to be used for parallel computation on
Linux and MacOS. Defaults to either \code{options("Ncpus")} if set, or 1
(single threaded) otherwise. Only single-threaded computation is currently
supported on Windows.}

\item{.warn_when_prop_few_obs}{Threshold proportion (0-1) for warning when
individuals have 0 or 1 observations. Defaults to 0.25 (i.e., warn when
25\% or more individuals have few observations).}

\item{...}{Additional arguments passed to the longitudinal model fitting
function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class \code{\link{LandmarkAnalysis}}.
}
\description{
Fits the specified longitudinal model for time-varying covariates up to
the landmark times
}
\details{
\subsection{Parallel processing}{

As the longitudinal model for each landmark time is independent of
the longitudinal models for other landmark times, parallel processing can
be used to vastly speed up computation. Currently only Unix-like operating systems
are supported by \code{landmaRk}.
}
}
\seealso{
\code{\link[lcmm:hlme]{lcmm::hlme()}} and \code{\link[lme4:lmer]{lme4::lmer()}} for additional arguments.
}
