% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{fit_survival,Landmarking-method}
\alias{fit_survival,Landmarking-method}
\title{Fits the specified survival model at the landmark times and up to the horizon
times specified by the user}
\usage{
\S4method{fit_survival}{Landmarking}(x, formula, landmarks, horizons, method, dynamic_covariates = c())
}
\value{
An object of class Landmarking.
}
\description{
Fits the specified survival model at the landmark times and up to the horizon
times specified by the user
}
\details{
\subsection{Mathematical formulation}{

This function estimates the conditional probability of survival to horizon
\eqn{s+w}, conditioned on having survived to the landmark time, \eqn{s},
that is
\deqn{\pi_i(s+w \vert s) = P(T_i > s+w \vert T_i \ge s, \bar{x}_i(s)), }
where \eqn{i} denotes an individual's index, \eqn{T_i} is the time to event
outcome for individual \eqn{i} and \eqn{\bar{x}_i(s)} are the covariates
observed for individual \eqn{i}, including the observed history of dynamic
covariates.

#' @inheritParams fit_survival
}
}
