% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks-lastfirst-yara.r
\name{get_landmarks_lastfirst_yara}
\alias{get_landmarks_lastfirst_yara}
\title{Firstlast-generated Landmarks}
\usage{
get_landmarks_lastfirst_yara(
  x,
  n = NULL,
  k = NULL,
  dist_method = "euclidean",
  seed_index = 1,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{n}{the number of landmarks requested.}

\item{k}{the desired cardinality of each cover set.}

\item{dist_method}{the distance metric to use. Any distance measure in the
\code{proxy} package is supported.}

\item{seed_index}{the first landmark to seed the algorithm.}

\item{shuffle_data}{whether to first randomly shuffle the data.}
}
\description{
Computes a landmark set based on nearest neighborhoods.
}
\details{
This function adapts the maxmin procedure to produce landmark points
dispersed according to the orders in which they are reached from each
other, rather than to their distances from each other. (Say more.)
}
