% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmaRk-package.R
\docType{package}
\name{landmaRk}
\alias{landmaRk-package}
\alias{landmaRk}
\title{landmaRk}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a modular end-to-end framework for dynamic risk prediction based on time-to-event and longitudinal data. This allows flexible specifications for the longitudinal and survival sub-models. The 'landmaRk' package enables reproducible benchmarks of different model choices, including cross-validation to assess out-of-sample predictive performance. Methods are described in Velasco-Pardo, et al. (2026) \href{https://arxiv.org/abs/2606.24678}{arXiv:2606.24678}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vallejosgroup.github.io/landmaRk/}
  \item \url{https://github.com/VallejosGroup/landmaRk}
  \item Report bugs at \url{https://github.com/VallejosGroup/landmaRk/issues}
}

}
\author{
\strong{Maintainer}: Victor Velasco-Pardo \email{vvelasco@ed.ac.uk} (\href{https://orcid.org/0000-0002-7166-1573}{ORCID})

Authors:
\itemize{
  \item Victor Velasco-Pardo \email{vvelasco@ed.ac.uk} (\href{https://orcid.org/0000-0002-7166-1573}{ORCID})
  \item Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk} (\href{https://orcid.org/0000-0002-4437-8713}{ORCID})
  \item Charlie Lees \email{charlie.lees@ed.ac.uk} (\href{https://orcid.org/0000-0002-0732-8215}{ORCID})
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk} (\href{https://orcid.org/0000-0003-3638-1960}{ORCID})
}

}
\keyword{internal}
