% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{landmark_maxmin}
\alias{landmark_maxmin}
\title{Maxmin in C++ (original)}
\usage{
landmark_maxmin(x, num, seed_index = 1L)
}
\arguments{
\item{x}{a data matrix}

\item{num}{desired number of landmark points, or number of sets, in a ball
cover (should be a positive integer)}

\item{seed_index}{index of the first landmark used to seed the algorithm}
}
\description{
Use the Euclidean maxmin procedure to choose landmarks

Maxmin procedure to choose 'num' landmarks for balls of fixed
radius (supports euclidean distance only)
}
\author{
Matt Piekenbrock
}
