% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/landmarks-lastfirst-cory.r,
%   R/landmarks-lastfirst-yara.r
\name{landmarks_lastfirst_cpp}
\alias{landmarks_lastfirst_cpp}
\alias{landmarks_lastfirst}
\alias{landmarks_lastfirst_cory}
\alias{landmarks_lastfirst_yara}
\title{Firstlast-generated Landmarks}
\usage{
landmarks_lastfirst_cpp(x, cardinality, seed_index = 0L)

landmarks_lastfirst_cory(
  x,
  number = NULL,
  cardinality = NULL,
  dist_method = "euclidean",
  seed_index = 1L,
  shuffle_data = FALSE
)

landmarks_lastfirst_yara(
  x,
  n = NULL,
  k = NULL,
  dist_method = "euclidean",
  seed_index = 1,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{cardinality}{the desired cardinality of each cover set.}

\item{seed_index}{the first landmark to seed the algorithm.}

\item{number}{the number of landmarks requested.}

\item{dist_method}{the distance metric to use. Any distance measure in the
\code{proxy} package is supported.}

\item{shuffle_data}{whether to first randomly shuffle the data.}

\item{n}{the number of landmarks requested.}

\item{k}{the desired cardinality of each cover set.}
}
\description{
Compute landmark points using maxmin procedure.

Computes a landmark set based on nearest neighborhoods.

Computes a landmark set based on nearest neighborhoods.
}
\details{
This function adapts the maxmin procedure to produce landmark points
dispersed according to the orders in which they are reached from each
other, rather than to their distances from each other. (Say more.)

This function adapts the maxmin procedure to produce landmark points
dispersed according to the orders in which they are reached from each
other, rather than to their distances from each other. (Say more.)
}
\author{
Jason Cory Brunson
}
