% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.r
\docType{package}
\name{maxmin}
\alias{maxmin}
\title{maxmin procedures to generate landmark sets for finite metric spaces}
\description{
Calculate landmark sets for finite metric spaces using the
maxmin procedure (for fixed-radius balls) or an adaptation of it for rank
data (for roughly fixed-cardinality nearest neighborhoods).
}
