% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance metrics}
\usage{
performance_metrics(
  x,
  landmarks,
  horizons,
  c_index = TRUE,
  brier = TRUE,
  auc_t = FALSE,
  train = TRUE,
  .k = 0
)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{horizons}{Vector of prediction horizons up to when the survival
submodel is fitted.}

\item{c_index}{A logical. If TRUE (default), C index is reported.}

\item{brier}{A logical. If TRUE (default), Brier score is reported.}

\item{auc_t}{A logical. If TRUE (default), AUC_t is reported.}

\item{train}{A logical. If TRUE (default), performance metrics are computed
in the training set. If FALSE (default), they are computed in the test set.}

\item{.k}{A numeric. Refers to the CV fold where in-sample and out-of-sample
prediction metrics.}
}
\value{
Data frame with performance metrics across the specified landmark
times and prediction horizons.
}
\description{
Computes concordance index (c-index) and Brier scores at the specified
landmark times and prediction horizons.
}
