% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_locf.R
\name{predict_locf_}
\alias{predict_locf_}
\title{Makes predictions from an LOCF process}
\usage{
predict_locf_(x, newdata, subject)
}
\arguments{
\item{x}{A dataframe}

\item{newdata}{A data frame containing static covariates and individual
IDs}

\item{subject}{Name of the column in newdata where individual IDs are stored.}
}
\value{
A vector of predictions.
}
\description{
Makes predictions from an LOCF process
}
