% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{predict_survival,LandmarkAnalysis-method}
\alias{predict_survival,LandmarkAnalysis-method}
\title{Make predictions for time-to-event outcomes at specified horizon times}
\usage{
\S4method{predict_survival}{LandmarkAnalysis}(x, landmarks, windows, method, ...)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{windows}{Vector of prediction windows determining horizon times.}

\item{method}{R function that is used to make predictions}

\item{...}{Additional arguments passed to the prediction function (e.g.
number of classes/clusters for lcmm).}
}
\value{
An object of class \code{\link{LandmarkAnalysis}}.
}
\description{
Make predictions for time-to-event outcomes at specified horizon times
}
