% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{prune,LandmarkAnalysis-method}
\alias{prune,LandmarkAnalysis-method}
\title{Prunes a landmark time from a \code{\link{LandmarkAnalysis}}, removing
the risk set, longitudinal submodel and survival submodel from the object.}
\usage{
\S4method{prune}{LandmarkAnalysis}(x, landmark = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{landmark}{A numeric indicating the landmark time.}
}
\value{
An object of class \code{\link{LandmarkAnalysis}}.
}
\description{
Prunes a landmark time from a \code{\link{LandmarkAnalysis}}, removing
the risk set, longitudinal submodel and survival submodel from the object.
}
