% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{prune}
\alias{prune}
\title{Prunes a landmark time from a \code{\link{LandmarkAnalysis}}, removing
the risk set, longitudinal submodel and survival submodel from the object.}
\usage{
prune(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{...}{Additional arguments}
}
\value{
An object of class \code{\link{LandmarkAnalysis}}.
}
\description{
Prunes a landmark time from a \code{\link{LandmarkAnalysis}}, removing
the risk set, longitudinal submodel and survival submodel from the object.
}
