% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{show}
\alias{show}
\title{Displays an object of class "\code{\link{LandmarkAnalysis}}"}
\usage{
show(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{...}{Additional arguments}
}
\description{
Displays an object of class "\code{\link{LandmarkAnalysis}}"
}
