% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{summary}
\alias{summary}
\title{Summarises landmarking model fits}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{...}{Additional arguments}
}
\value{
A summary of the desired submodel
}
\description{
Summarises landmarking model fits
}
