% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose-list.r
\name{transpose_list}
\alias{transpose_list}
\title{Convert between lists of members and lists of memberships}
\usage{
transpose_list(x, sparse = TRUE)
}
\arguments{
\item{x}{A list of positive integer vectors. Names will be ignored.}

\item{sparse}{Logical; whether to use a slower \code{lapply()} implementation
rather than a speedier but possibly memory-intensive implementation that
constructs the element-set matrix. (Neither implementation currentnly takes
advantage of sparse matrix objects.)}
}
\description{
Covers may be stored as lists of set members or as lists of
element memberships. This function converts one to the other (the operation
is an involution).
}
\examples{
# transpose an example list (note that an index is missing)
lst <- list(1:3,3:5,c(3,7))
transpose_list(lst)
# observe that transposition is an involution
transpose_list(transpose_list(lst))
}
