% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_lcmm.R
\name{check_lcmm_convergence}
\alias{check_lcmm_convergence}
\title{Checks convergence of lcmm models}
\usage{
check_lcmm_convergence(x)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}
}
\description{
Checks convergence of lcmm models
}
