% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{fit_longitudinal}
\alias{fit_longitudinal}
\title{Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times}
\usage{
fit_longitudinal(
  x,
  landmarks,
  method,
  formula,
  dynamic_covariates,
  cores = getOption("Ncpus", 1L),
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{A vector of Landmark times.}

\item{method}{Either "lcmm" or "lme4" or a function for fitting a
longitudinal data model, where the first argument is a formula, and also
has a \code{data} argument.}

\item{formula}{A formula to be used in longitudinal sub-model fitting.}

\item{dynamic_covariates}{Vector of time-varying covariates to be modelled
as the outcome of a longitudinal model.}

\item{cores}{Number of cores/threads to be used for parallel computation.
Defaults to either \code{options("Ncpus")} if set, or 1 (single threaded)
otherwise.}

\item{...}{Additional arguments passed to the longitudinal model fitting
function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class \code{\link{Landmarking}}.
}
\description{
Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times
}
\seealso{
\code{\link[lcmm:hlme]{lcmm::hlme()}} and \code{\link[lme4:lmer]{lme4::lmer()}} for additional arguments.
}
