% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{fit_survival,Landmarking-method}
\alias{fit_survival,Landmarking-method}
\title{Fits the specified survival model at the landmark times and up to the horizon
times specified by the user}
\usage{
\S4method{fit_survival}{Landmarking}(x, formula, landmarks, horizons, method, dynamic_covariates = c())
}
\value{
An object of class Landmarking.
}
\description{
#' @inheritParams fit_survival
}
