% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark.r
\docType{package}
\name{landmark}
\alias{landmark}
\title{procedures to generate landmark sets from finite metric spaces}
\description{
Calculate landmark sets for finite metric spaces using the
maxmin procedure (for fixed-radius balls) or an adaptation of it for rank
data (for roughly fixed-cardinality nearest neighborhoods).
}
\author{
Matt Piekenbrock

Jason Cory Brunson

Yara Skaf
}
