% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/landmarks-lastfirst.r
\name{landmarks_lastfirst_cpp}
\alias{landmarks_lastfirst_cpp}
\alias{landmarks_lastfirst}
\alias{firstlast_R}
\alias{landmarks_lastfirst_R}
\title{Neighborhood-based Landmark Sets}
\usage{
landmarks_lastfirst_cpp(x, num_sets = 0L, cardinality = 0L, seed_index = 0L)

firstlast_R(x, dist_method = "euclidean", ties_method = "min")

landmarks_lastfirst_R(
  x,
  dist_method = "euclidean",
  ties_method = "min",
  pick_method = "first",
  num_sets = NULL,
  cardinality = NULL,
  seed_index = 1L,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{num_sets}{a positive integer; the desired number of landmark points, or
of sets in a neighborhood cover.}

\item{cardinality}{a positive integer; the desired cardinality of each
landmark neighborhood, or of each set in a landmark cover.}

\item{seed_index}{an integer (the first landmark to seed the algorithm) or
one of the character strings \code{"random"} (to select a seed uniformly at
random) and \code{"firstlast"} (to select a seed from the firstlast set).}

\item{dist_method}{a character string specifying the distance metric to use;
passed to \code{proxy::dist(method)}. Any distance measure in the \code{proxy}
package is supported.}

\item{ties_method}{a character string specifying the method for handling
ties; passed to \code{rank(ties.method)}. Only \code{"min"} has been tested and is
recommended.}

\item{pick_method}{a character string specifying the method for selecting
from indistinguishable points, either \code{"first"} (the default) or
\code{"random"}.}

\item{shuffle_data}{logical; whether to randomly shuffle the data before
running a procedure.}
}
\description{
Compute landmark points using maxmin procedure.

Compute landmark sets based on nearest neighborhoods.
}
\details{
These functions adapt the maxmin procedure to produce landmark
points dispersed according to the orders in which they are reached from
each other, rather than to their distances from each other. (Say more.)

One, both, or neither of \code{num_sets} and \code{cardinality} may be passed values.
If neither is specified, then \code{num_sets} is defaulted to \code{24L}. If the
values yield neighborhoods that do not cover \code{x}, then, effectively,
\code{num_sets} is increased until the cardinality necessary to cover \code{x} is at
most \code{cardinality}. To generte a complete landmark set, use \code{cardinality = 1L}.
}
\author{
Jason Cory Brunson

Yara Skaf
}
