% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/landmarks-maxmin.r
\name{landmarks_maxmin_cpp}
\alias{landmarks_maxmin_cpp}
\alias{landmark_maxmin}
\alias{landmarks_maxmin}
\alias{minmax_R}
\alias{chebyshev_center_R}
\alias{landmarks_maxmin_R}
\title{Ball-based Landmark Sets}
\usage{
landmarks_maxmin_cpp(x, num_sets = 0L, radius = 0, seed_index = 0L)

landmark_maxmin(x, n, seed_index = 0L)

minmax_R(x, dist_method = "euclidean")

chebyshev_center_R(x, dist_method = "euclidean")

landmarks_maxmin(
  x,
  dist_method = "euclidean",
  num_sets = NULL,
  radius = NULL,
  seed_index = 1L,
  shuffle_data = FALSE
)

landmarks_maxmin_R(
  x,
  dist_method = "euclidean",
  pick_method = "first",
  num_sets = NULL,
  radius = NULL,
  seed_index = 1L
)
}
\arguments{
\item{x}{a data matrix.}

\item{num_sets}{a positive integer; the desired number of landmark points, or
of sets in a ball cover.}

\item{radius}{a positive number; the desired radius of each
landmark ball, or of each set in a ball cover.}

\item{seed_index}{an integer (the first landmark to seed the algorithm) or
one of the character strings \code{"random"} (to select a seed uniformly at
random) and \code{"minmax"} (to select a seed from the minmax set).}

\item{dist_method}{a character string specifying the distance metric to use;
passed to \code{proxy::dist(method)}. Any distance measure in the \code{proxy}
package is supported.}

\item{shuffle_data}{whether to first randomly shuffle the data.}

\item{pick_method}{a character string specifying the method for selecting
from indistinguishable points, either \code{"first"} (the default) or
\code{"random"}.}
}
\description{
Compute landmark points using maxmin procedure.

Compute landmark sets based on fixed-radius balls.
}
\details{
This function uses the maxmin procedure to produce a set of evenly
spaced landmark points from a data set. Maxmin is a simple greedy
algorithm that is relatively efficient, but it has a tendency to pick out
extremal points.

One, both, or neither of \code{num_sets} and \code{radius} may be passed values. If
neither is specified, then \code{radius} is internally set to \code{0} so that a
complete landmark set is generated. If the values yield balls that do not
cover \code{x}, then their number is increased until the radius necessary to
cover \code{x} is at most \code{radius}.
}
\references{
De Silva, Vin, and Gunnar E. Carlsson. "Topological estimation
using witness complexes." SPBG 4 (2004): 157-166.

Dłotko, Paweł. "Ball Mapper: A Shape Summary for Topological Data
Analysis." (2019). Web.
}
\author{
Matt Piekenbrock

Jason Cory Brunson
}
