% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/landmarks-maxmin.r
\name{landmarks_maxmin_cpp}
\alias{landmarks_maxmin_cpp}
\alias{landmark_maxmin}
\alias{landmarks_maxmin}
\alias{minmax}
\alias{minmax_R}
\alias{chebyshev_center_R}
\alias{landmarks_maxmin_orig}
\alias{landmarks_maxmin_R}
\title{Ball-based Landmark Sets}
\usage{
landmarks_maxmin_cpp(x, num_sets = 0L, radius = -1, seed_index = 0L)

landmark_maxmin(x, n, seed_index = 1L)

minmax(x, dist_method = "euclidean")

minmax_R(x, dist_method = "euclidean")

chebyshev_center_R(x, dist_method = "euclidean")

landmarks_maxmin(
  x,
  dist_method = "euclidean",
  pick_method = "first",
  num_sets = NULL,
  radius = NULL,
  frac = FALSE,
  seed_index = 1L,
  cover = FALSE,
  engine = NULL
)

landmarks_maxmin_orig(
  x,
  dist_method = "euclidean",
  pick_method = "first",
  num_sets = NULL,
  radius = NULL,
  seed_index = 1L
)

landmarks_maxmin_R(
  x,
  dist_method = "euclidean",
  pick_method = "first",
  num_sets = NULL,
  radius = NULL,
  frac = FALSE,
  seed_index = 1L,
  cover = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{num_sets}{a positive integer; the desired number of landmark points, or
of sets in a ball cover.}

\item{radius}{a positive number; the desired radius of each
landmark ball, or of each set in a ball cover.}

\item{seed_index}{an integer (the first landmark to seed the algorithm) or
one of the character strings \code{"random"} (to select a seed uniformly at
random) and \code{"minmax"} (to select a seed from the minmax set).}

\item{dist_method}{a character string specifying the distance metric to use;
passed to \code{proxy::dist(method)}. Any distance measure in the \code{proxy}
package is supported.}

\item{pick_method}{a character string specifying the method for selecting one
among indistinguishable points, either \code{"first"} (the default), \code{"last"},
or \code{"random"}.}

\item{frac}{logical; whether to treat \code{radius} as a fraction of the diameter
of \code{x}.}

\item{cover}{logical; whether to return a data frame of landmark indices and
cover sets (by member index) rather than only a vector of landmark indices.}
}
\description{
Compute landmark points using maxmin procedure.

Compute landmark sets based on fixed-radius balls.
}
\details{
This function uses the maxmin procedure to produce a set of evenly
spaced landmark points from a data set. Maxmin is a simple greedy algorithm
that is relatively efficient, but it has a tendency to pick out extremal
points.

One, both, or neither of \code{num_sets} and \code{radius} may be passed values. If
neither is specified, then \code{num_sets} is defaulted to \code{24L}. If the values
yield balls that do not cover \code{x}, then their number is increased until the
radius necessary to cover \code{x} is at most \code{radius}. To generte a complete
landmark set, use \code{radius = 0L}.
}
\references{
De Silva, Vin, and Gunnar E. Carlsson. "Topological estimation
using witness complexes." SPBG 4 (2004): 157-166.

Dłotko, Paweł. "Ball Mapper: A Shape Summary for Topological Data
Analysis." (2019). Web.
}
\author{
Matt Piekenbrock

Jason Cory Brunson

Yara Skaf
}
