% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/landmarks-maxmin.r
\name{landmark_maxmin}
\alias{landmark_maxmin}
\alias{landmarks_maxmin}
\title{Maxmin-generated Landmarks}
\usage{
landmark_maxmin(x, n, seed = 0L)

landmarks_maxmin(
  x,
  n = NULL,
  eps = NULL,
  dist_method = "euclidean",
  seed_index = 1,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{n}{the number of landmarks requested.}

\item{eps}{the desired radius of balls in the cover set.}

\item{dist_method}{the distance metric to use. Any distance measure in the
\code{proxy} package is supported.}

\item{seed_index}{the first landmark to seed the algorithm.}

\item{shuffle_data}{whether to first randomly shuffle the data.}
}
\description{
Computes a set of \code{n} landmarks of a point cloud.
}
\details{
This function uses the greedy maxmin procedure to produce a set of
\code{n} evenly spaced landmark points in the data set \code{x}. Maxmin is
a simple greedy algorithm that is relatively efficient, but it has a
tendency to pick out extremal points. If the distance metric is euclidean,
an efficient Rcpp implementation is used. If another metric is requested,
the algorithm is performed in R. Alternatively, a radius \eqn{\epsilon}
can be specified so that the landmark set is computed to cover the space
with \eqn{\epsilon}-balls centered at the landmarks.
}
\references{
De Silva, Vin, and Gunnar E. Carlsson. "Topological estimation
using witness complexes." SPBG 4 (2004): 157-166.

Dłotko, Paweł. "Ball Mapper: A Shape Summary for Topological Data
Analysis." (2019). Web.
}
