% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{landmarks_maxmin_cpp}
\alias{landmarks_maxmin_cpp}
\title{Maxmin in C++ (extended)}
\usage{
landmarks_maxmin_cpp(x, num = 0L, radius = -1, seed_index = 1L, cover = FALSE)
}
\arguments{
\item{x}{a data matrix}

\item{num}{desired number of landmark points, or number of sets, in a ball
cover (should be a positive integer)}

\item{radius}{desired radius of a cover set (should be a positive real number)}

\item{seed_index}{index of the first landmark used to seed the algorithm}

\item{cover}{boolean specifying whether to return cover sets in addition to
the landmark points}
}
\description{
Maxmin procedure to choose landmarks for balls of fixed radius
(supports Euclidean distance only)
}
\author{
Jason Cory Brunson

Yara Skaf
}
