% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks-neighborhood-cory.r
\name{landmarks_neighborhood}
\alias{landmarks_neighborhood}
\alias{chebycenter_firstlast_cory}
\alias{landmarks_lastfirst_cory}
\title{Neighborhood-based Landmark Sets}
\usage{
chebycenter_firstlast_cory(x, y = NULL, dist_method = "euclidean")

landmarks_lastfirst_cory(
  x,
  number = NULL,
  cardinality = NULL,
  dist_method = "euclidean",
  seed_index = 1L,
  shuffle_data = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{y}{a data matrix, or \code{NULL} to recycle \code{x}.}

\item{dist_method}{the distance metric to use. Any distance measure in the
\code{proxy} package is supported.}

\item{number}{the number of landmarks requested.}

\item{cardinality}{the desired cardinality of each cover set.}

\item{seed_index}{the first landmark to seed the algorithm.}

\item{shuffle_data}{whether to first randomly shuffle the data.}
}
\description{
Compute landmark sets based on nearest neighborhoods.
}
\details{
These functions adapt the maxmin procedure to produce landmark
points dispersed according to the orders in which they are reached from
each other, rather than to their distances from each other. (Say more.)

One, both, or neither of \code{number} and \code{cardinality} may be passed values.
If neither is specified, then \code{cardinality} is internally set to \code{1L} so
that a complete landmark set is generated. If the values yield
neighborhoods that do not cover \code{x}, then, effectively, \code{number} is
increased until the cardinality necessary to cover \code{x} is at most
\code{cardinality}.
}
\author{
Jason Cory Brunson
}
