% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_visualisation.R
\name{plot,Landmarking-method}
\alias{plot,Landmarking-method}
\title{Plots survival curves for the fitted landmarking models.}
\usage{
\S4method{plot}{Landmarking}(x, type = "survival", ...)
}
\arguments{
\item{Landmarking}{An object of class Landmarking.}
}
\description{
Plots survival curves for the fitted landmarking models.
}
