% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{predict_survival}
\alias{predict_survival}
\title{Make predictions for time-to-event outcomes at specified horizon times}
\usage{
predict_survival(
  x,
  landmarks,
  horizons,
  method,
  dynamic_covariates = c(),
  include_clusters = FALSE,
  validation_fold = 0,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{horizons}{Vector of prediction horizons up to when the survival submodel
is fitted.}

\item{method}{R function that is used to make predictions}

\item{dynamic_covariates}{Vector of time-varying covariates to be used
in the survival model.}

\item{include_clusters}{Boolean indicating whether to propagate cluster
membership to survival analysis.}

\item{validation_fold}{If positive, cross-validation fold where model is
fitted. If 0 (default), model fitting is performed on the complete dataset.}

\item{...}{Additional arguments passed to the prediction function (e.g.
number of classes/clusters for lcmm).}
}
\value{
An object of class \code{\link{LandmarkAnalysis}}.
}
\description{
Make predictions for time-to-event outcomes at specified horizon times
}
