% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{show,LandmarkAnalysis-method}
\alias{show,LandmarkAnalysis-method}
\title{Displays an object of class "\code{\link{LandmarkAnalysis}}"}
\usage{
\S4method{show}{LandmarkAnalysis}(object)
}
\arguments{
\item{object}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{...}{Additional arguments}
}
\description{
Displays an object of class "\code{\link{LandmarkAnalysis}}"
}
