% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{summary,LandmarkAnalysis-method}
\alias{summary,LandmarkAnalysis-method}
\title{Summarise a LandmarkAnalysis object}
\usage{
\S4method{summary}{LandmarkAnalysis}(
  object,
  type = c("longitudinal", "survival"),
  landmark,
  horizon = NULL,
  dynamic_covariate = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{type}{If \code{longitudinal}, it summarises the longitudinal submodel.
If \code{survival}, it summarises the survival submodel.}

\item{landmark}{A numeric indicating the landmark time.}

\item{horizon}{For survival submodels, a numeric indicating the horizon time.}

\item{dynamic_covariate}{For longitudinal submodels, a character indicating
the dynamic covariate}
}
\value{
A summary of the desired submodel, printed to the console.
Returns \code{NULL} invisibly.
}
\description{
Summarise a LandmarkAnalysis object
}
\examples{
data(epileptic)
epileptic_dfs <- split_wide_df(
  epileptic,
  ids = "id", times = "time",
  static = c("with.time", "with.status", "treat", "age", "gender", "learn.dis"),
  dynamic = c("dose"),
  measurement_name = "value"
)
x <- LandmarkAnalysis(
  data_static = epileptic_dfs$df_static,
  data_dynamic = epileptic_dfs$df_dynamic,
  event_indicator = "with.status",
  ids = "id", event_time = "with.time",
  times = "time", measurements = "value"
) |>
  compute_risk_sets(365.25) |>
  fit_survival(
    formula = survival::Surv(event_time, event_status) ~ treat + age,
    landmarks = 365.25,
    horizons = 2 * 365.25,
    method = "coxph"
  )
summary(x, type = "survival", landmark = 365.25, horizon = 2 * 365.25)
\donttest{
# Full pipeline including longitudinal submodel
x2 <- LandmarkAnalysis(
  data_static = epileptic_dfs$df_static,
  data_dynamic = epileptic_dfs$df_dynamic,
  event_indicator = "with.status",
  ids = "id", event_time = "with.time",
  times = "time", measurements = "value"
) |>
  compute_risk_sets(365.25) |>
  fit_longitudinal(
    landmarks = 365.25,
    method = "lme4",
    formula = value ~ treat + age + gender + learn.dis + (time | id),
    dynamic_covariates = c("dose")
  ) |>
  predict_longitudinal(
    landmarks = 365.25,
    method = "lme4",
    allow.new.levels = TRUE,
    dynamic_covariates = c("dose")
  ) |>
  fit_survival(
    formula = survival::Surv(event_time, event_status) ~
      treat + age + gender + learn.dis + dose,
    landmarks = 365.25,
    horizons = 2 * 365.25,
    method = "coxph",
    dynamic_covariates = c("dose")
  ) |>
  predict_survival(landmarks = 365.25, horizons = 2 * 365.25)
summary(x2, type = "longitudinal", landmark = 365.25, dynamic_covariate = "dose")
summary(x2, type = "survival", landmark = 365.25, horizon = 2 * 365.25)
}
}
