% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{summary}
\alias{summary}
\title{Summarises landmarking model fits}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{LandmarkAnalysis}}.}

\item{type}{If \code{longitudinal}, it summarises the longitudinal submodel.
If \code{survival}, it summarises the survival submodel.}

\item{landmark}{A numeric indicating the landmark time.}

\item{horizon}{For survival submodels, a numeric indicating the horizon time.}

\item{dynamic_covariate}{For longitudinal submodels, a character indicating
the dynamic covariate}
}
\value{
A summary of the desired submodel
}
\description{
Summarises landmarking model fits
}
